[Experimental] Correctly handles showing saves in xEdit mode, up to conflict coloring in view pane.
Synced to r1843
Required fix in wbFiles when calling TwbFileSource.Create

Adds 3 properties to handle the Tree pane. TreeHead, TreeLeaf and TreeBranch.
Properly display the coSaves tree now. Some definition work still to do for the save tree.

Compare to... functional for saves. Including a change in the default directories to use depending on type of file.
Added save extensions to the Comapre To Open dialog.



git-svn-id: http://skyrim-plugin-decoding-project.googlecode.com/svn@1844 bdfc8fb3-84cf-a68a-e902-ed0415cf5604